Add Gregorian Calendar class
This commit adds a time point class which follows the name in the kata
GregorianCalendar. The name is misleading, but keeping it for the sake
of correspondence.

Added the basic constructor shown in the kata, but added also type safe
constructors.

Basic unit tests follow the implementation. Additional ones shall be put
in place to cover additional edge cases. For demonstration purposes, it
is believed the number of UT is sufficient.