Add Workday Calendar class
This commit adds the basic features for a workday calendar. By
definining a start and stop time of a workday and holidays we can
calculate the end date providing an increment as float.

At the moment it supports clamping of start and stop time as well as
correction of start and stop date to business working hours. Some edge
cases are to be handled but the current state is able to correctly
calculate the Kata results.

Unit tests follow and a parameterized test calculates all the kata
scenarios and compares it against the expected values.